feat(reducers/searchstatereducer): moved reducer logic to separate file
