+++ src/pages/index.tsx
@@ -1,5 +1,4 @@
-import { Dispatch, SetStateAction, useReducer } from 'react';
-import { actionType } from '@/components/SearchConfigButton';
+import { Dispatch, SetStateAction } from 'react';
 import { Flex } from '@chakra-ui/react';
 import SearchConfigCard from '@/layouts/SearchConfigCard';
 import { Layout, PageHeading, Button } from '@/components/CustomComponents';
@@ -9,54 +8,11 @@ import { User } from 'next-auth';
 import { Ticket, useTicket } from '@/contexts/ticketContext';
 import { NextRouter, useRouter } from 'next/router';
 import endpoint from '../../endpoints.config';
-
-type searchConfigType = {
-  roleSelection: Record<string, boolean>;
-  serverSelection: Record<string, boolean>;
-};
-
-const defaultSearchConfig = {
-  roleSelection: {
-    rockie: false,
-    coach: false,
-  },
-  serverSelection: {
-    us: false,
-    eu: false,
-    sea: false,
-  },
-};
-
-const roleSelectionDictionary = ['Player', 'Coach'];
-const serverSelectionDictionary = ['US', 'EU', 'SEA'];
-const toggleSearchConfig = (
-  state: searchConfigType,
-  action: actionType
-): searchConfigType => {
-  const { configValue } = action;
-  switch (action.configType) {
-    case 'role':
-      return {
-        ...state,
-        roleSelection: {
-          ...state.roleSelection,
-          [configValue]: !state.roleSelection[configValue],
-        },
-      };
-      break;
-    case 'server':
-      return {
-        ...state,
-        serverSelection: {
-          ...state.serverSelection,
-          [configValue]: !state.serverSelection[configValue],
-        },
-      };
-      break;
-    default:
-      throw new Error('Unexcepted Action Type');
-  }
-};
+import {
+  useSearchConfig,
+  roleSelectionDictionary,
+  serverSelectionDictionary,
+} from '../reducers/searchStateReducer';
 
 const postData = async ({ id, steamID }: User): Promise<Ticket> => {
   const ticketID = await fetch(`${endpoint.NEXT_PUBLIC_API_RMM}/ticket`, {
@@ -83,26 +39,25 @@ const createTicket = async (
   router.push('/searching');
 };
 function Index({ session }: { session: Session }): React.ReactElement {
-  const [
-    { roleSelection, serverSelection },
-    dispatchSearchConfigState,
-  ] = useReducer(toggleSearchConfig, defaultSearchConfig);
   const { setTicket } = useTicket();
   const router = useRouter();
-
+  const [
+    { roleSelection, serverSelection },
+    dispatchSearchConfig,
+  ] = useSearchConfig();
   return (
     <Layout>
       <PageHeading>Are You Ready ?</PageHeading>
       <Flex w={['20rem', '36rem']} justifyContent="space-around">
         <SearchConfigCard
-          dispatchSearchConfigState={dispatchSearchConfigState}
+          dispatchSearchConfig={dispatchSearchConfig}
           dictionary={roleSelectionDictionary}
           configState={roleSelection}
           configType="role"
           configTitle="Roles"
         />
         <SearchConfigCard
-          dispatchSearchConfigState={dispatchSearchConfigState}
+          dispatchSearchConfig={dispatchSearchConfig}
           dictionary={serverSelectionDictionary}
           configState={serverSelection}
           configType="server"
